Fix compilation errors

+++ src/avr/disassembler.rs
@@ -1,3 +1,3 @@
 pub trait Disassembler {
-  fn disassemble(&self, address: u16) -> (String, Option<String>, Option<String>);
+  fn disassemble(&self, address: u32) -> (String, Option<String>, Option<String>);
 }
