+++ src/avr/disassemblers/bset.rs
@@ -4,7 +4,7 @@ use crate::avr::instructions::bset::Bset;
 impl Disassembler for Bset {
   fn disassemble(
     &self,
-    _address: u16,
+    _address: u32,
   ) -> (
     std::string::String,
     Option<std::string::String>,
