+++ src/avr/disassemblers/icall22.rs
@@ -4,7 +4,7 @@ use crate::avr::instructions::icall22::Icall22;
 impl Disassembler for Icall22 {
   fn disassemble(
     &self,
-    _address: u16,
+    _address: u32,
   ) -> (
     std::string::String,
     Option<std::string::String>,
