+++ src/avr/disassemblers/muls.rs
@@ -4,7 +4,7 @@ use crate::avr::instructions::muls::Muls;
 impl Disassembler for Muls {
   fn disassemble(
     &self,
-    _address: u16,
+    _address: u32,
   ) -> (
     std::string::String,
     Option<std::string::String>,
