+++ src/avr/disassemblers/reti16.rs
@@ -4,7 +4,7 @@ use crate::avr::instructions::reti16::Reti16;
 impl Disassembler for Reti16 {
   fn disassemble(
     &self,
-    _address: u16,
+    _address: u32,
   ) -> (
     std::string::String,
     Option<std::string::String>,
